Move get_valid_objects function to collective.nitf.upgrades
